+++ cpp/subprojects/boosting/src/boosting/output/predictor_regression_label_wise.cpp
@@ -21,7 +21,7 @@ namespace boosting {
 
         #pragma omp parallel for firstprivate(numExamples) firstprivate(rulePtr) firstprivate(featureMatrixPtr) \
         firstprivate(predictionMatrixPtr) schedule(dynamic) num_threads(numThreads_)
-        for (uint32 i = 0; i < numExamples; i++) {
+        for (intp i = 0; i < numExamples; i++) {
             applyRule(*rulePtr, featureMatrixPtr->row_cbegin(i), featureMatrixPtr->row_cend(i),
                       predictionMatrixPtr->row_begin(i));
         }
@@ -38,7 +38,7 @@ namespace boosting {
 
         #pragma omp parallel for firstprivate(numExamples) firstprivate(rulePtr) firstprivate(featureMatrixPtr) \
         firstprivate(predictionMatrixPtr) schedule(dynamic) num_threads(numThreads_)
-        for (uint32 i = 0; i < numExamples; i++) {
+        for (intp i = 0; i < numExamples; i++) {
             float32 tmpArray1[numFeatures];
             uint32 tmpArray2[numFeatures] = {};
             applyRuleCsr(*rulePtr, featureMatrixPtr->row_indices_cbegin(i), featureMatrixPtr->row_indices_cend(i),
@@ -57,7 +57,7 @@ namespace boosting {
 
         #pragma omp parallel for firstprivate(numExamples) firstprivate(modelPtr) firstprivate(featureMatrixPtr) \
         firstprivate(predictionMatrixPtr) schedule(dynamic) num_threads(numThreads_)
-        for (uint32 i = 0; i < numExamples; i++) {
+        for (intp i = 0; i < numExamples; i++) {
             for (auto it = modelPtr->used_cbegin(); it != modelPtr->used_cend(); it++) {
                 const Rule& rule = *it;
                 applyRule(rule, featureMatrixPtr->row_cbegin(i), featureMatrixPtr->row_cend(i),
@@ -77,7 +77,7 @@ namespace boosting {
 
         #pragma omp parallel for firstprivate(numExamples) firstprivate(modelPtr) firstprivate(featureMatrixPtr) \
         firstprivate(predictionMatrixPtr) schedule(dynamic) num_threads(numThreads_)
-        for (uint32 i = 0; i < numExamples; i++) {
+        for (intp i = 0; i < numExamples; i++) {
             float32 tmpArray1[numFeatures];
             uint32 tmpArray2[numFeatures] = {};
             uint32 n = 1;
